requirements.txt
# API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv>=1.0.0
neo4j==5.17.0
google-generativeai==0.3.1
mistralai==0.0.7
openai==1.3.5
requests==2.31.0
httpx>=0.24.1
tiktoken==0.5.2
supabase>=2.3.0

# Logging
rich==13.7.0
loguru==0.7.2
python-json-logger==2.0.7
tenacity==8.2.3

# Development
black>=23.12.1
flake8>=7.0.0